:art: core插件迁移完成
