--- build-logic/core/debug/build.gradle.kts
@@ -1,23 +1,0 @@
-plugins {
-    `kotlin-dsl`
-}
-java.toolchain.languageVersion.set(JavaLanguageVersion.of(11))
-
-dependencies {
-
-    implementation(project(":core:base"))
-
-}
-
-gradlePlugin {
-
-    plugins {
-        create("module.debug") {
-
-            implementationClass = "ModuleDebugPlugin"
-            id = "module.debug"
-
-        }
-    }
-
-}
